Move rider collection methods into AE system data models
