+++ module/data/active-effect/_module.mjs
@@ -7,6 +7,7 @@ export {
   ConditionData,
   EnchantmentData
 };
+export {default as RiderProviderMixin} from "./mixins/rider-provider.mjs";
 
 export const config = {
   base: BaseEffectData,
